feat(tests): Try to fix test REST req

+++ tests/integration/__init__.py
@@ -19,13 +19,13 @@ class TestLineTcpSender(unittest.TestCase):
     def receive(self, table_name):
         received = False
         while not received:
-            time.sleep(0.25)
+            time.sleep(0.1)
             r = req.get(
                 url=self.URL, params={"query": f"select * from '{table_name}';"}
             )
             data = json.loads(r.text)
-            print(data)
-            if data["count"] > 0:
+
+            if "error" not in data and data["count"] > 0:
                 req.get(url=self.URL, params={"query": f"drop table '{table_name}';"})
                 received = True
